Updating per springdo's suggestions. Still working on adding JSON validation, but getting this out there so the new versioned api routes are visible.
