+++ package.json
@@ -3,10 +3,12 @@
   "version": "0.1.0",
   "description": "Open Management Portal Backend",
   "main": "src/index.js",
-  "repository": "git@github.com:logandonley/omp-be.git",
+  "repository": "git@github.com:rht-labs/omp-be.git",
   "license": "Apache",
   "dependencies": {
+    "body-parser": "^1.18.3",
     "express": "^4.16.4",
+    "express-validator": "^5.3.0",
     "mongodb": "^3.1.9",
     "mongoose": "^5.3.10"
   },
