+++ src/controllers/v1/cluster.controller.js
@@ -1,18 +1,54 @@
-const Group = require("../models/group.model.js");
+const Cluster = require("../../models/v1/cluster.model.js");
+const { checkSchema } = require("express-validator/check");
 
 exports.create = (req, res) => {
-  if (!req.body) {
-    return res.status(400).send({
-      message: "Group content can't be empty"
-    });
+  // const errors = validationResult(req);
+  // if (!errors.isEmpty()) {
+  //   return res.status(422).json({ errors: errors.array() });
+  // }
+  let size = {};
+  switch (req.body.size) {
+    case "SMALL":
+      size = {
+        master_node: 1,
+        infra_node: 1,
+        app_node: 2
+      };
+      break;
+    case "MEDIUM":
+      size = {
+        master_node: 1,
+        infra_node: 1,
+        app_node: 3
+      };
+      break;
+    case "LARGE":
+      size = {
+        master_node: 1,
+        infra_node: 1,
+        app_node: 5
+      };
+      break;
+    default:
+      size = {
+        master_node: 1,
+        infra_node: 1,
+        app_node: 3
+      };
+      break;
   }
-  const group = new Group({
-    group_name: req.body.group_name.toLowerCase(),
-    display_name: req.body.display_name,
-    role: req.body.role.toLowerCase()
+  const cluster = new Cluster({
+    env_id: req.body.env_id,
+    customer_id: req.body.customer_id,
+    logging: req.body.logging,
+    metrics: req.body.metrics,
+    size: size,
+    ha: req.body.ha,
+    ocp_version: req.body.ocp_version,
+    hosting_platform: req.body.hosting_platform
   });
 
-  group
+  cluster
     .save()
     .then(data => {
       res.send(data);
@@ -25,7 +61,7 @@ exports.create = (req, res) => {
 };
 
 exports.findAll = (req, res) => {
-  Group.find({})
+  Cluster.find({})
     .then(data => {
       res.send(data);
     })
@@ -37,13 +73,13 @@ exports.findAll = (req, res) => {
 };
 
 exports.findOne = (req, res) => {
-  if (!req.params.group_name) {
+  if (!req.params.env_id) {
     res.status(400).send({
       message: "You must provide an environment id parameter."
     });
     return;
   }
-  Group.find({ group_name: req.params.group_name.toLowerCase() })
+  Cluster.find({ env_id: req.params.env_id.toLowerCase() })
     .then(data => {
       res.send(data);
     })
@@ -55,13 +91,13 @@ exports.findOne = (req, res) => {
 };
 
 exports.update = (req, res) => {
-  if (!req.params.group_name) {
+  if (!req.params.env_id) {
     res.status(400).send({
       message: "You must provide an environment id parameter."
     });
     return;
   }
-  Group.updateOne({ group_name: req.params.group_name.toLowerCase() }, req.body)
+  Cluster.updateOne({ env_id: req.params.env_id.toLowerCase() }, req.body)
     .then(data => {
       res.send(data);
     })
@@ -73,13 +109,13 @@ exports.update = (req, res) => {
 };
 
 exports.delete = (req, res) => {
-  if (!req.params.group_name) {
+  if (!req.params.env_id) {
     res.status(400).send({
       message: "You must provide an environment id parameter."
     });
     return;
   }
-  Group.deleteOne({ group_name: req.params.group_name.toLowerCase() })
+  Cluster.deleteOne({ env_id: req.params.env_id.toLowerCase() })
     .then(data => {
       res.send(data);
     })
