+++ src/routes/v1/prefix.js
@@ -1,0 +1,1 @@
+module.exports = "/api/v1";
